Improve route layout for syndicate manager (#26)
Currently, we have manager actions under syndicate folder with the route
resolution being /syndicate/manage.

Instead of having this as seperate route component, we are moving it to
/syndicates/[syndicateAddress]/manage route to make the syndicates route
the parent route. The manage component will be accessed from the syndicateAddress
obtained from the clicked syndicate on My Syndicates page. This will become handy when
working on manage actions