[ENG 526] [ENG-922] Add Change Syndicate Settings modal on manage page (#103)
Implement functionality for Manager settings modal.

Using this option, a manager can update/edit features of a syndicate. Some syndicate parameters can only be edited while the syndicate is open, others when the syndicate is closed and others are dependent on whether the syndicate is modifiable or not. All these quards have been handled.