[ENG-4387] App not detecting MetaMask extension (#1701)
When user clicks the MetaMask to connect their wallet, we need to detect if the
extension is installed and connect to it. If the extension is not installed, we
navigate the user to the download page from where they can install the extension.

Currently, the app is failing to detect if the extension is installed for cases where only metamask
is the only extension that is available. This commit fixes this issue so that we can
detect the extension in both cases.