Fix infinite loading state (#1702)
* Temporarily save details of a newly created club for use o portfolio.

Initially, we were saving limited club information during creation in which we ended
up making contract calls to the protocol. This had a result of some the calls failing
because the contracts had not fully synced which was resulting to an infinite loading
state on the portfolio page. The new changes aim to temporarily save all the information
that we need for a given club and then remove them once the graph has synced with the
new data. This eliminates the need to make contract calls to the protocol.

* add check for currentMintPolicy for new clubs

this adds an additional case where redux state for new clubs might be out of sync. in which case makes several contract calls to get club details with guardMixinManager + and sets currentMintPolicy to match guardMixinManager address

* [ENG-4335] /3 conditional check for mintPolicy address for goerli + polygon

since not all mintPolicies are deployed on each chain conditonally calls contract if the address is available.

Co-authored-by: nikki-kiga <42276368+nikki-kiga@users.noreply.github.com>