[ENG-4041] Hide tokens that are non-transferable. (#1728)
* [ENG-4041] Hide tokens that are non-transferable.

There are token, for example, IC tokens which can not be transferred to members when making distributions.
For a better user experience, these tokens need not be displayed to the user/admin. This ticket serves to
find those tokens and remove them from token selection step. Tokens manually hidden by the admin are also excluded from asset list.