Fix ETH ICs not working (#1860)
I noticed that when I create an IC accepting ETH as deposit, the details page shows that the club
is accepting USDC. The root cause is that for ETH, depositToken is set to the zero address and thus our
check for this fails and a subsequent call to graph is called with the zero address as the token address,
which eventually fails. The fix is to include the check whether depositToken is zero address.

Also this commit has a few type fixes for CONTRACT_ADDRESS object and the way its being accessed across the app.